[refer]: Regression-test Savannah #68681.
* src/preproc/refer/tests/check-index-file-validity.sh: Do it.

+++ src/preproc/refer/tests/check-index-file-validity.sh
@@ -90,6 +90,22 @@ printf '\377\377\377\377' | dd of="$REFER".i bs=1 seek=24 conv=notrunc
 echo "checking behavior when index file parameters corrupt" >&2
 "$lkbib" Schemer 2>&1 > /dev/null | grep -q 'error.*corrupt' || wail
 
+# Regression-test Savannah #68681.
+
+REFER="$sandbox_dir"/string-pool-test.bib
+echo '%A Test' > "$REFER"
+
+if ! "$indxbib" -c "$common_words_file" "$REFER"
+then
+    echo "cannot generate index for bibliography file; skipping" >&2
+    exit 77 # skip
+fi
+
+# Corrupt the beginning of the string pool.  Thanks to Pavol Sloboda.
+printf '\0' | dd of="$REFER".i bs=1 seek=4044 conv=notrunc
+echo "checking behavior when index file string pool corrupt" >&2
+"$lkbib" Test 2>&1 > /dev/null | grep -q 'error.*corrupt string' || wail
+
 # We need `-f` because "little-schemer.bib" gets mode 444 in a "make
 # distcheck" build.
 rm -rf "$sandbox_dir"
